+++ LibretroRT.FrontendComponents.Win2DRenderer/Win2DRenderer.cs
@@ -89,7 +89,7 @@ namespace LibretroRT.FrontendComponents.Win2DRenderer
                         return false;
                     }
 
-                    GameID = mainGameFilePath.SHA1();
+                    GameID = mainGameFilePath;
                     RenderTargetManager.CurrentCorePixelFormat = core.PixelFormat;
                     CoreIsExecuting = true;
                     return true;
