+++ RetriX.Shared/Services/ISaveStateService.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 
 namespace RetriX.Shared.Services
 {
     public interface ISaveStateService
     {
-        string GameId { get; set; }
+        void SetGameId(string id);
 
         Task<byte[]> LoadStateAsync(uint slotId);
         Task<bool> SaveStateAsync(uint slotId, byte[] data);
